Change how we compare values slightly
Instead of comparing the config hash, we create a *new* object from the updating
json file, then compare both by looping through the list of attributes that can
be updated according to marathon.
We then return the object and use *that* object instead to update the marathon
app in marathon.