Version needs to not be checked
Because it will be different each time: the versions set by marathon are
time-stamped, so they increment every time and there is no way for you to
know.
GPUs for some reason are always set to None, even tho the API returns 0 if
it's not set. This trips the "needs update" flag so setting to 0 to avoid
an extra trip to the API.