Allow JSON output and multiple apps per file
This is to support dump and restore functionality: now you can use
``--list-apps --json`` and that will produce a big JSON structure with all the
apps, you can then redirect to a file and that is now your backup.
Also modified ``--config-file`` so it can support a single app definition or
multiple app definitions, this way it can keep working as it is right now but
can also support dump files like those produced with the above command.